inheriting from a struct fails when the file is loaded twice
for some totally weird reason, starting chef-server via the rackup file
causes mixlib/authentication to get loaded twice. This causes a
superclass mismatch when inheriting from Struct.new. Since we aren't
adding any behavior to the struct, we can just assign it to a constant
and get the same results.