Fix following bug:
It's possible that a request contains more than nine headers like "x_ops_authorization_n".
In this case headers will be sorted in the wrong way. The first will be "x_ops_authorization_1",
the second "x_ops_authorization_10" and so on. So that request signature transfered by parts in "x_ops_authorization_n" headers will be reconstructed in wrong way. So that authentication will fail.