Fix SignedObject::DEFAULT_SERVER_API_VERSION constant scope
The DEFAULT_SERVER_API_VERSION is not a constant of the SignedHeaderAuth
module. Rather than forcing the constant lookup in that module we'll
remove the namespace and allow the Ruby constant lookup to handle it for
us. This resolves a bug where the constant lookup fails when
mixlib-authentication is vendored into other projects.

Signed-off-by: Ryan Cragun <me@ryan.ec>